Allow multiple html tags to be scraped at the same time
It makes sense to scrape details like the author or the date of the
article on different times, since they're are going to be rearragned
anyway when constructing the final markdown file, but if one wants to
include links, subtitles and images from the article in the markdown
file, those obviously have to be scraped at the same time as the text to
preserve the order of them. Therefore, this update allow the content
part of the profiles to contain more kinds of elements than just one
(which would normally be <p>), so that it's possible in one function to
scrape all the bulk content of an article.