Add error catch for non existent html for details
Sometimes details from the article like the author or date might not be
available, and the script would crash since its trying to extract HTML
elements that doesn't exist. This update adds an error catch for that,
and returns "Unknown" should the detail that script is looking for not
be there. Also adds it so the script raise exception if trying to fetch
the main text return unknown.