Add function for scrambling the OG tags
Since this script tries to emulate a feed, the OG tags should be
scrambled so that the "feed" isn't seperated in different news site, but
with articles from different sources mixed. The reason for doing it this
way instead of simply scrambling the article links before scraping the
OG Tags is that this preserves the order of the articles when it comes
to when they were released, so an older article will never appear before
a newer one from the same source.