Add the name of news site to title
When giving an overview of the articles it would be benificial to have
the name of the news site behind the article in the title of the
article. Some news site already has this, some doesn't. This commit
makes the script add the name of the news site to the title on those who
doesn't already have it.