Add function for scraping the articles after overview
In the overview the user will be presented with a lot of articles, and
when one is choosen the browser will navigate to that article and the
user can read it. When the user has read it and closes the article, the
script will read the source code and parse that to the markdown file.
Still needs to find a way to parse the profilename though, as the only
thing that gets parsed through the webdriver is the article itself.