Add communication of profile
Since the only communication that happened between the first and second
part of the script on each side of scraping the article was the url of
the article, it was prior not possible to communicate about what profile
was associated with the article. This updates however adds the
profilename as a paramter to all the article links, which can then
later be scraped and parsed and the name of the profile for the article
the user chooses can thereby be extracted.