Disable recursive scraping of article body to avoid duplicates
What would sometimes happen was that when scraping for multiple tags in
the body (like both the p and img tag) there would fx. appear an img tag
inside a parent p tag, which would lead the script to srape the img tag
two times, first as a child of the p tag and then as an img tag which
would create quite a lot of duplicates. To combat this the recursive
function of the function locateContent has been disabled when scraping
for the body of the article, which will make it only scrape the prior
duplicate tags once instead of twice, this time only as it's own tag and
not as a child. Will need to specify all the wanted tags to be scraped
in the profiles from now on though, as it is no more possible to just
rely on the elements one wants to be scraped to be a child of an element
specified in the profile.