Check if articleOverview needs refresh
Refreshing the articleOverview is by far the part of the script that
takes longest to complete, but if the overview was refreshed just 2
minutes ago, it doesn't make sense to refresh it again. This update
implements a feature that automatically skips updating the article
overview if it has been updated within the last 2 hours.