Add error catch for when the script can extract OSINTwProfile
The method utilized for detecting when the user closed the selenium
browser - which means the script can then continue on to scrape the page -
unfortunatly isn't very stable and sometimes registers that the user
closed the page even if that wasn't the case and then throws an error
when trying to scrape the wrong page since its activating to early. For
the time being a fix doesn't seem possible as the selenium browser
doesn't have an option in the api for detecting when the browser is
closed.